+++ container-engine-lib/docs/changelog.md
@@ -2,6 +2,9 @@
 ### Features
 * Added module CRUD methods
 
+### Fixes
+* Fixed a bug in `CreateAPIContainer` where the wrong label was being used to check for enclave ID
+
 # 0.10.2
 ### Features
 * Added `UserService` methods, `Modules` methods and `CreateRepartition` method in `KurtosisBackend` interface